tests/wpt/web-platform-tests/html/semantics/document-metadata/the-meta-element/pragma-directives/attr-meta-http-equiv-refresh/not-in-shadow-tree.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Meta refresh only applies while in the document tree, not in a shadow tree</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://html.spec.whatwg.org/multipage/semantics.html#pragma-directives">

<div id="log"></div>
<script>
"use strict";

const iframe = document.createElement("iframe");
iframe.src = "support/ufoo";

let loadCount = 0;

iframe.onload = () => {
  ++loadCount;
  const iDocument = iframe.contentDocument;

  if (loadCount === 1) {
    const div = iDocument.createElement("div");
    assert_true('attachShadow' in div, 'attachShadow support');
    const shadowRoot = div.attachShadow({ mode: "open" });
    shadowRoot.innerHTML = `<meta http-equiv="refresh" content="1; url=foo">`;
    iDocument.body.appendChild(div);

    // Want to make sure no refreshes happen
    step_timeout(done, 3000);
  } else {
    assert_unreached("Got more than 1 load event");
  }
};

document.body.appendChild(iframe);
</script>
